Fix millis() rollover handling across shade movement and connection timing
Two classes of bug, both surfacing only around the ~49.7-day millis()
wraparound:

- Several deadlines (gpioRelease, lastFrame.await, tx_queue.delay_time,
  rebootDelay.rebootTime, WiFi/MQTT/GitOTA connect timers, the login
  lockout, the SSDP send queue) were computed as `millis() + duration`
  and checked with a direct `millis() > deadline` comparison, which
  breaks for one deadline-sized window every time millis() wraps.
  Replaced with the rollover-safe `(int32_t)(millis() - deadline) >= 0`
  form throughout.

- moveStart/tiltStart/sunStart/windStart/noSunStart/noWindStart/windLast
  (Somfy.h) and MQTT's lastConnect were uint64_t, even though millis()
  only ever produces 32 bits of range before wrapping. Their comparisons
  already used the "safe" `curTime - start >= duration` subtraction
  pattern, but doing that subtraction in 64-bit arithmetic doesn't
  cancel out the wrap the same way a 32-bit subtraction does, so the
  safety was illusory. Narrowed these fields (and the curTime locals
  that feed them) to uint32_t so the subtraction wraps at the same
  modulus as millis() itself.

Also fixed rebootDelay_t::rebootTime, which was a plain signed int
instead of a type matched to millis().

+++ src/MQTT.h
@@ -9,7 +9,7 @@ private:
   const char* makeTopic(const char* topic);
 
 public:
-  uint64_t lastConnect = 0;
+  uint32_t lastConnect = 0;
   bool suspended = false;
   char clientId[32] = {'\0'};
